// ui/hooks/generated/types.ts
export interface TransportOptions {
  id: string;
  dtlsParameters?: DtlsParameters | null;
  iceCandidates: IceCandidate[];
  iceParameters?: IceParameters | null;
}

export interface MediaListResponse {
  media: MediaResponse[];
}

export interface DateTime {
  timestamp: number;
}

export interface ServerCommitResponse {
  voiceId: string;
  commitId: string;
  accepted: boolean;
  errorMessage: string;
}
// ...
export interface ErrorResponse {
  errorMessage: string;
  errorCode: string;
}

export interface ChatMessage {
}

export interface DtlsParameters {
  fingerprints: DtlsFingerprint[];
  role?: string | null;
}


export type VoiceRequest =
| { type: "init"; data: InitRequest }
| { type: "connectProducerTransport"; data: ConnectTransportRequest }
| { type: "produce"; data: ProduceRequest }
| { type: "closeProducer"; data: CloseProducerRequest }
| { type: "connectConsumerTransport"; data: ConnectTransportRequest }
| { type: "consume"; data: ConsumeRequest }
| { type: "consumerResume"; data: ConsumerResumeRequest }
| { type: "voiceData"; data: VoiceDataRequest }
| { type: "clientCommit"; data: ClientCommitRequest }
| { type: "commitAck"; data: CommitAckRequest }
;

export interface InitResponse {
  roomId: string;
  participantId: string;
  consumerTransportOptions?: TransportOptions | null;
  producerTransportOptions?: TransportOptions | null;
  routerRtpCapabilities?: RtpCapabilitiesFinalized | null;
}

export interface ConnectedTransportResponse {
  success: boolean;
}

export interface ConsumerResumeRequest {
  consumerId: string;
}

export interface MediaResponse {
  media_id: string;
  filename: string;
  timestamp: number;
}

export interface VoiceDataNotification {
  userId: number;
  voiceId: string;
  data: number[];
}

export interface RtpCapabilities {
  codecs: RtpCodecCapability[];
  headerExtensions: RtpHeaderExtension[];
}

export interface CommitAckRequest {
  commitId: string;
  success: boolean;
  errorMessage: string;
}

export interface RtpParameters {
  mid?: string | null;
  codecs: RtpCodecParameters[];
  headerExtensions: RtpHeaderExtensionParameters[];
  encodings: RtpEncodingParameters[];
  rtcp?: RtcpParameters | null;
  msid?: string | null;
}


export type JoinMode = "Open" | "InviteOnly" | "RequestToJoin";

export interface RtpCapabilitiesFinalized {
  codecs: RtpCodecCapabilityFinalized[];
  headerExtensions: RtpHeaderExtension[];
}

export interface ConnectTransportRequest {
  dtlsParameters?: DtlsParameters | null;
}

export interface CloseProducerRequest {
  producerId: string;
}

export interface ConsumeRequest {
  producerId: string;
}
// ...
export interface ProducerRemovedNotification {
  participantId: string;
  producerId: string;
}
// ...
export type VoiceResponse =
| { type: "init"; data: InitResponse }
| { type: "producerAdded"; data: ProducerAddedNotification }
| { type: "producerRemoved"; data: ProducerRemovedNotification }
| { type: "connectedProducerTransport"; data: ConnectedTransportResponse }
| { type: "produced"; data: ProducedResponse }
| { type: "connectedConsumerTransport"; data: ConnectedTransportResponse }
| { type: "consumed"; data: ConsumedResponse }
| { type: "error"; data: ErrorResponse }
| { type: "voiceData"; data: VoiceDataNotification }
| { type: "serverCommit"; data: ServerCommitNotification }
| { type: "addProposal"; data: AddProposalNotification }
| { type: "serverCommitResponse"; data: ServerCommitResponse }
// ...
export interface ClientCommitRequest {
  voiceId: string;
  commit: number[];
  welcomeMessage: number[];
}

export interface RtpRtx {
  ssrc: number;
}

export interface RtpCodecParameters {
  mimeType: string;
  payloadType: number;
  clockRate: number;
  channels?: number | null;
  parameters: Record<string, string>;
  rtcpFeedback: RtcpFeedback[];
}

export interface ConsumedResponse {
  consumerId: string;
  producerId: string;
  kind: string;
  rtpParameters?: RtpParameters | null;
}

export interface AddProposalNotification {
  voiceId: string;
  proposal: number[];
}
// ...
export interface InitRequest {
  roomId: string;
  rtpCapabilities?: RtpCapabilities | null;
}

export interface UpdateUserAvatarResponse {
  success: boolean;
  avatar_url: string;
}

export interface ServerCommitNotification {
  voiceId: string;
  commit: number[];
  commitId: string;
}

export interface DisplayUserInfo {
  user_id: number;
  username: string;
  avatar: string;
  status: string;
  last_seen: number;
  created_at: number;
  trust_level: number;
}

export interface RtpHeaderExtensionParameters {
  uri: string;
  id: number;
  encrypt: boolean;
  parameters: Record<string, string>;
}

export interface ProducerAddedNotification {
  participantId: string;
  producerId: string;
  appData: string;
}

export interface ProducedResponse {
  producerId: string;
}
// ...
export interface VoiceDataRequest {
  voiceId: string;
  data: number[];
}
// ...
export interface ProduceRequest {
  kind: string;
  rtpParameters?: RtpParameters | null;
  appData: string;
}